router improvements + break out slow test
